Break up Fields unit tests
